ResponseDto operations completed for Category

+++ Core/CafeAPI.Application/Services/Concretes/CategoryService.cs
@@ -15,10 +15,18 @@ public class CategoryService : ICategoryService
         _categoryRepository = categoryRepository;
         _mapper = mapper;
     }
-    public async Task AddCategoryAsync(CreateCategoryDto categoryDto)
+    public async Task<ResponseDto<object>> AddCategoryAsync(CreateCategoryDto categoryDto)
     {
-        var category = _mapper.Map<Category>(categoryDto);
-        await _categoryRepository.AddAsync(category);
+        try
+        {
+            var category = _mapper.Map<Category>(categoryDto);
+            await _categoryRepository.AddAsync(category);
+            return new ResponseDto<object> { Success = true, Data = category, Message = $"{category.Name} başarı ile oluştruldu" };
+        }
+        catch (Exception)
+        {
+            return new ResponseDto<object> { Success = false, Message = "Bir Hata Oluştu", ErrorCodes = ErrorCodes.Exception };
+        }
     }
     public async Task<ResponseDto<List<ResultCategoryDto>>> GetAllCategoriesAsync()
     {
@@ -30,25 +38,55 @@ public class CategoryService : ICategoryService
             var result = _mapper.Map<List<ResultCategoryDto>>(categories);
             return new ResponseDto<List<ResultCategoryDto>> { Success = true, Data = result };
         }
-        catch (Exception ex)
+        catch (Exception)
         {
-            return new ResponseDto<List<ResultCategoryDto>> { Success = false, Message = ex.Message, ErrorCodes = ErrorCodes.Exception };
+            return new ResponseDto<List<ResultCategoryDto>> { Success = false, Message = "Bir Hata Oluştu", ErrorCodes = ErrorCodes.Exception };
         }
     }
-    public async Task<DetailCategoryDto> GetByIdCategoryAsync(int id)
+    public async Task<ResponseDto<DetailCategoryDto>> GetByIdCategoryAsync(int id)
     {
-        var category = await _categoryRepository.GetByIdAsync(id);
-        var result = _mapper.Map<DetailCategoryDto>(category);
-        return result;
+        try
+        {
+            var category = await _categoryRepository.GetByIdAsync(id);
+            if (category is null)
+                return new ResponseDto<DetailCategoryDto> { Success = false, Message = "Kategori Bulunamadı", ErrorCodes = ErrorCodes.NotFound };
+            var result = _mapper.Map<DetailCategoryDto>(category);
+            return new ResponseDto<DetailCategoryDto> { Success = true, Data = result};
+        }
+        catch (Exception)
+        {
+            return new ResponseDto<DetailCategoryDto> { Success = false, Message = "Bir Hata Oluştu", ErrorCodes = ErrorCodes.Exception };
+        }
     }
-    public async Task RemoveCategoryAsync(int id)
+    public async Task<ResponseDto<object>> RemoveCategoryAsync(int id)
     {
-        var category = await _categoryRepository.GetByIdAsync(id);
-        await _categoryRepository.RemoveAsync(category);
+        try
+        {
+            var category = await _categoryRepository.GetByIdAsync(id);
+            if (category is null)
+                return new ResponseDto<object> { Success = false, Message = "Kategori Bulunamadı", ErrorCodes = ErrorCodes.NotFound };
+            await _categoryRepository.RemoveAsync(category);
+            return new ResponseDto<object> { Success = true, Message = $"{category.Name} isimli Kategori Silindi"};
+        }
+        catch (Exception)
+        {
+            return new ResponseDto<object> { Success = false, Message = "Bir Hata Oluştu", ErrorCodes = ErrorCodes.Exception };
+        }
     }
-    public async Task UpdateCategoryAsync(UpdateCategoryDto categoryDto)
+    public async Task<ResponseDto<object>> UpdateCategoryAsync(UpdateCategoryDto categoryDto)
     {
-        var category = _mapper.Map<Category>(categoryDto);
-        await _categoryRepository.UpdateAsync(category);
+        try
+        {
+            var result = await _categoryRepository.GetByIdAsync(categoryDto.Id);
+            if (result is null)
+                return new ResponseDto<object> { Success = false, Message = "Kategori Bulunamadı", ErrorCodes = ErrorCodes.NotFound };
+            var category = _mapper.Map(categoryDto,result);
+            await _categoryRepository.UpdateAsync(category);
+            return new ResponseDto<object> { Success = true, Message = $"{category.Name} isimli kategori başarı ile güncellendi" };
+        }
+        catch (Exception)
+        {
+            return new ResponseDto<object> { Success = false, Message = "Bir Hata Oluştu", ErrorCodes = ErrorCodes.Exception};
+        }
     }
 }
